requirements.txt
torch
torchvision
torchaudio
diffusers
wandb
pytorch_lightning
omegaconf
scipy
scikit-image
scikit-learn
accelerate
opencv-python
opencv-contrib-python
einops
timm
albumentations
deepspeed
gradio
onnx
onnxruntime-gpu
tensorrt-cu12